Fix ZodError
Closes #96

+++ src/utils/github-fs.ts
@@ -2,7 +2,12 @@ import { Buffer } from "buffer"
 import mime from "mime"
 import { GitHubRepository } from "../types"
 
-const GITHUB_ENDPOINT = "https://api.github.com"
+function getFileEndpoint(githubRepo: GitHubRepository, path: string) {
+  return `https://api.github.com/repos/${githubRepo.owner}/${githubRepo.name}/contents/${
+    // Remove leading slash if present
+    path.replace(/^\//, "")
+  }`
+}
 
 type GetFileShaOptions = {
   githubToken: string
@@ -15,13 +20,8 @@ export async function getFileSha({
   githubRepo,
   path,
 }: GetFileShaOptions): Promise<string | null> {
-  const endpoint = `${GITHUB_ENDPOINT}/repos/${githubRepo.owner}/${githubRepo.name}/contents/${
-    // Remove leading slash if present
-    path.replace(/^\//, "")
-  }`
-
   // Get the SHA of the file
-  const response = await fetch(endpoint, {
+  const response = await fetch(getFileEndpoint(githubRepo, path), {
     headers: {
       Accept: "application/vnd.github.v3+json",
       Authorization: `Bearer ${githubToken}`,
@@ -47,21 +47,51 @@ type ReadFileOptions = {
 }
 
 export async function readFile({ githubToken, githubRepo, path }: ReadFileOptions) {
-  const response = await fetch(
-    `https://api.github.com/repos/${githubRepo.owner}/${githubRepo.name}/contents/${
-      // Remove leading slash if present
-      path.replace(/^\//, "")
-    }`,
-    {
-      headers: {
-        Accept: "application/vnd.github.raw",
-        Authorization: `Bearer ${githubToken}`,
-      },
+  const response = await fetch(getFileEndpoint(githubRepo, path), {
+    headers: {
+      Accept: "application/vnd.github.v3+json",
+      Authorization: `Bearer ${githubToken}`,
     },
-  )
+  })
+
+  if (!response.ok) {
+    console.error(response)
+
+    switch (response.status) {
+      // Unauthorized
+      case 401:
+        throw new Error(`Invalid GitHub token`)
+
+      // Not found
+      case 404:
+        throw new Error(`File not found: ${path} in ${githubRepo.owner}/${githubRepo.name}`)
+
+      // Other error
+      default:
+        throw new Error(
+          `Failed to fetch file: ${path} in ${githubRepo.owner}/${githubRepo.name} (${response.status})`,
+        )
+    }
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const { content } = (await response.json()) as any
+
+  return Buffer.from(content, "base64").toString("utf8")
+}
+
+// Used to read large files like images or PDFs
+export async function readRawFile({ githubToken, githubRepo, path }: ReadFileOptions) {
+  const response = await fetch(getFileEndpoint(githubRepo, path), {
+    headers: {
+      Accept: "application/vnd.github.raw",
+      Authorization: `Bearer ${githubToken}`,
+    },
+  })
 
   if (!response.ok || !response.body) {
     console.error(response)
+
     switch (response.status) {
       // Unauthorized
       case 401:
@@ -117,17 +147,14 @@ type WriteFileOptions = {
 }
 
 export async function writeFile({ githubToken, githubRepo, path, content }: WriteFileOptions) {
-  const endpoint = `${GITHUB_ENDPOINT}/repos/${githubRepo.owner}/${githubRepo.name}/contents/${
-    // Remove leading slash if present
-    path.replace(/^\//, "")
-  }`
-
   // Get the SHA of the file
   const sha = await getFileSha({ githubToken, githubRepo, path })
+
+  // If the file doesn't exist, `sha` will be null
   const fileExists = Boolean(sha)
 
   // Create or update the file
-  const response = await fetch(endpoint, {
+  const response = await fetch(getFileEndpoint(githubRepo, path), {
     method: "PUT",
     headers: {
       Authorization: `Bearer ${githubToken}`,
@@ -142,6 +169,7 @@ export async function writeFile({ githubToken, githubRepo, path, content }: Writ
 
   if (!response.ok) {
     console.error(response)
+
     throw new Error(
       `Failed to ${fileExists ? "update" : "create"} file: ${path} ${response.status}}`,
     )
@@ -155,16 +183,16 @@ type DeleteFileOptions = {
 }
 
 export async function deleteFile({ githubToken, githubRepo, path }: DeleteFileOptions) {
-  const endpoint = `${GITHUB_ENDPOINT}/repos/${githubRepo.owner}/${githubRepo.name}/contents/${
-    // Remove leading slash if present
-    path.replace(/^\//, "")
-  }`
-
   // Get the SHA of the file
   const sha = await getFileSha({ githubToken, githubRepo, path })
 
+  if (!sha) {
+    // File doesn't exist, so we don't need to delete it
+    return
+  }
+
   // Delete the file
-  const response = await fetch(endpoint, {
+  const response = await fetch(getFileEndpoint(githubRepo, path), {
     method: "DELETE",
     headers: {
       Authorization: `Bearer ${githubToken}`,
@@ -177,6 +205,7 @@ export async function deleteFile({ githubToken, githubRepo, path }: DeleteFileOp
 
   if (!response.ok) {
     console.error(response)
+
     throw new Error(`Failed to delete file: ${path} (${response.status})`)
   }
 }
